backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
torch
torchvision
torchaudio
pillow
numpy
pandas
scikit-learn
openai
google-generativeai
python-dotenv
google-auth
requests
os